Initial split into modular components
